safety

+++ src/s21_string.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-unsigned long s21_strlen(const char *str);
+size_t s21_strlen(const char *str);
 int s21_strcmp(const char *str1, const char *str2);
 char *s21_strcpy(char *dest, const char *src);
 char *s21_strcat(char *dest, const char *src);
